add documentation for CKEditor plugin architecture
this adds some documentation showing how to add a custom plugin to
CKEditor, complete with code examples, explanations, and a lot of links
to the CKEditor documentation.

this documentation will also be updated soon to include information
about how to extend our Markdown plugin, once that API is merged.

part of #39
pr #87