add CKEditor media embed plugin
this adds the CKEditor MediaEmbed plugin to our CKEditor setup, which
adds support for embedding various types of media directly into the
editor.

I also added a setup for extending the editor's markdown syntax support
as part of a CKEditor plugin, by writing a MarkdownSyntaxPlugin that
defines rules for translating a new node type to and from markdown. This
should make it easy for us to add custom node types in the future, and
to co-locate the markdown code with the CKEditor plugin code.

Documentation will be added to our CKEditor docs soon to explain how
this all works.

for #39
pr #99