add a minimal configuration of CKEditor
This adds a minimal configuratin of CKEditor which can be used to edit
Markdown but without all of the rich options we want in the full course
content authoring UI.

Basically, now you can pass a minimal boolean prop to the MarkdownEditor
component. When this is passed it will enable fewer editing features
(i.e. no image or media insertion, no headers).

This is also added this to the test page
(http://localhost:8043/markdown-editor) with a tabbed UI with the full
editor in one tab and the minimal editor in another.

closes #111
pr #112