fix turndown bug relating to <li> tags
this fixes an issue with how we were overriding turndown's built-in rule
for <li> tags. Basically, there is an issue with the default rule that
turndown uses to translate an <li> into Markdown (see
https://github.com/domchristie/turndown/issues/291 for details).

We were overriding this behavior with the behavior we want by adding a
new rule for <li>, but this was causing an issue because now we had two
different rules that were both matching <li> tags.

Anyhow, this commit switches it to instead update the rule in place, so
we still only have one rule, fixing an issue that came up with the
editor.

closes #113
pr #114