fix double-instantiation issue w/ ckeditor
this fixes an issue relating to adding turndown rules - basically, when
the editor is instantiated and destroyed more than once while the
application is running we need to have a way to ensure that we only set
turndown rules once, lest we run into issues. I added a little check to
make sure this isn't done.

closes #142
pr #144