Add validation for multiple select, min / max
This adds validation to our form validation code for the multiple-select
variant of our select widget (meaning when a `WidgetVariant.Select`
field has the `multiple` property set to `true`).

In addition to validating that the value for that field makes sense,
this commit also adds validation for the `min` and `max` property, which
can be used to set an expected range for the number of selected values.

closes #185
pr #211