add relation field widget
This adds a new field type, WidgetVariant.Relation, which basically
implments a one-to-many relation between two collections.

See the google doc for details on the idea behind this and how it works.

Implementing this added a requirement to access the Website object
fairly deep in the component tree. Rather than pass that object down the
whole component tree I added a React Context (called WebsiteContext)
which lets us set a Website as the 'focused' website for a given
component tree. Using the `useContext` hook it's easy to get this value
out.

The Relation Widget itself basically just is some wrapper code around
our existing Select widget. All it does is fetch the records for
collection being linked to and format those into `Option` objects which can
be passed into the Relation widget. This approach lets us re-use the
existing initialization and validation code already written for the
Select widget for the Relation widget. Additionally, data for linked
content is saved in the same format as the Relation widget already saves
it (i.e. a list when `multiple === true` and as a single value when it
is `false`).

This should provide us with a good basis to build filtering on top of
the Relation widget, which will let us specify a field that links to
only one type of resource, for instance.

closes #241
pr #263