node-sass -> sass
this just removes the node-sass package and replaces it with the pure-JS
version of dart-sass, the newer and more up-to-date reference
implementatin of Sass.

background: node-sass depends on the libsass C library, which has now
been deprecated: https://sass-lang.com/blog/libsass-is-deprecated

dart-sass is a newer implementation in Dart which is more up to date
with the spec. Since Dart can be compiled to plain JavaScript, it is
also possible to avoid having a native extension by using it over
node-sass (the libsass bindings are a frequent cause of install
trouble).

pr #244