split out types for ConfigField
Currently we have a sort of all-in-one type on the frontend called
ConfigField which types the configuration that is supplied from the
backend for each field in the content form. This is sub-optimal, because
we actually have several different variants of these objects which have
different fields on them, which may be distinguished by the value they
have for `widget: WidgetVariant`.

Currently every single possible field is defined as an optional field on
the `ConfigField` type, which means that we have to do a few superfluous
null checks.

This commit splits the `ConfigField` type apart into several different
types like `MarkdownConfigField`, `FileConfigField`, `TextConfigField`,
etc. Wherever possible we can use this to be more specific about what
fields are going to be present on an object.

closes #278
pr #280