Add support for filtering in the relation widget
this adds support for filtering the pieces of content shown in the UI
for the relation widget, so that a restriction can be put in place as to
which entries are valid to link to. This can be used to, for instance,
implement a file type restriction on a link to a file-like piece of
content.

closes #289
pr #296