Implement linking to content in other websites
This allows the relation field to link to content in other websites, not
just in the same website.

In the future we plan to support three use-cases for the relation field:

- linking to content in a specific collection in the same website
- linking to content in a specific collection in a specific, other
  website
- linking to content in a user-specified website and collection (i.e.
  one specified when the content author is editing the content).

The first use-case was already supported, and this commit enables
support for the second. There will be a follow-up PR to add support for
the third.

The 2nd and 3rd use-cases require changing the data storage format.
Instead of just storing an array of UUIDs for linked-to content, we also
need to store which website they live in. So this commit makes that
change as well.

part of #288
pr #306