switch site content modal over to ModalState
In #378 I added a module called `modal_state.ts` which provides a nice
type-safe way to manage the state of a modal. This commit sets us up to
use that library for the site content modals, in addition to the
WebsiteCollection modals for which it was introduced.

`modal_state.ts` defines a type, ModalState, which can be one of three
states: Editing, Adding, and Closed. This captures the three UI states
our modals can be in, i.e. they are either closed, open to add new
content, or open to edit existing content.

The type is defined like this:

type ModalState<T> = Editing<T> | Adding<T> | Closed<T>

This is a bit like a Data type in Haskell or an Enum in Rust (think
Option or Result if you've written some Rust before) but it isn't quite
as clean because TypeScript doesn't have built-in support for this.

The type parameter <T> is the type of a value we need to correctly
render the Editing state. This could be, for instance, the uuid for the
object we want to edit. In the case of the site content in this PR it is
a string, for the `.text_id` property which we use as a uuid for site
content.

To capture that, this commit also introduces a 'concrete' type:

type WebsiteContentModalState = ModalState<string>

I made changes throughout the site content component tree to use a value
of this type for controlling both whether the modal was open or closed
and whether to show the adding or the editing UI.

pr #437