add resource widget plugin for CKEditor
This PR adds the first part of the resource embed functionality that we
want to add to our CKEditor-based Markdown editor.

Basically, this just defines a new node type in the editor (called a
SiteContentEmbed) and adds some basic placeholder stuff for serializing
this to and from a hugo shortcode ({{< resource $UUID >}}) and to a very
barebones display in the editor. There is no UI for adding these to the
editor yet, and no rich display when they are embedded in the editor.
This helps get some boilerplate out of the way so that subsequent PRs
don't have to be as large.

closes #452
pr #450