add resource embed UI
This adds the first part of the resource embed UI: an interface for
selecting resources and inserting them into the Markdown editor. This
works by hitting the site content API to get resources, displaying them
in a list, and then running a CKEditor insert command to add the
resource embed to the editor.

This does not include display a rich preview of those resources in the
editor itself.

There is a configuration change here. The Markdown widget now has an
optional 'attach' field, which holds the name of the resource which
should be inserted. In the default case this should be set to
`"resource"`.

closes #453
pr #475