fix drawer width bug
this PR fixes an issue where the drawer width wasn't always being set
correctly, because we weren't being so consistent about where in the
component hierarchy we were calling the addDefaultFields function on the
fields in a collection. To fix it, I basically hoisted that call up to
the SiteContentListing component, refactoring the function a bit in the
process, so that we have one consistent value throughout that part of
the component tree.

see issue for more details about the bug

closes #459
pr #460