Upgrade the ckeditor packages and webpack
This upgrades all the @ckeditor packages to 29.1.0 -- I'm determined to
stay up-to-date with these because I think it will be bad if we don't!

This also upgrades webpack from v4 to v5, following the guide here:
https://webpack.js.org/migrate/5/ This lets us also bump up the versions
of the various webpack-related packages (the loaders and whatnot) and
ensure we don't have a bunch of outdated dependencies.

pr #523