add rich display for videos embedded in markdown editor
This adds a basic display of a video resource when it has been added to
the Markdown editor. Since videos are going to be just uploaded to
YouTube, we can simply use a normal YouTube embed iframe to show a
'preview' of the video in the editor, alongside its title and
description.

closes #506
pr #544