Add support for linking to resources in the Markdown editor
This adds basic support for adding links to resources (not embeds) in
the Markdown editor. These are rendered to markdown via the
resource_link shortcode. In order to support this I added a new CKEditor
node type plugin and a new React component to actually render the things
when they're added to the editor. I also made some modifications to the
resource picker to allow the user to decide whether they want to embed
or link to a resource, after selecting a particular resource.

closes #541
closes #542
pr #560