allow users to create inline links to resources
This introduces a new dependency on our custom resource link plugin,
which is based on the code for CKEditor's standard link plugin. This
allows resource links to be inlined in the text rather than being block
nodes, by storing the UUID for the link as an attribute on a range of
text.

Resource links are translated to the `resource_link` shortcode, like
this:

```
{{< resource_link uuidforaresource "text inside the link" >}}
```

This change also required some changes to how we open the resource
picker and some related things. Now instead of having one 'Add resource'
button we now have 'Embed resource' and 'Link resource' buttons.
Together these make it a little less ambiguous for the user (at least,
that is the hope!)

closes #607
closes #612
part of #536