remove media embed plugin button from Markdown editor toolbar
this removes the button for the CKEditor Media Embed plugin
(https://ckeditor.com/docs/ckeditor5/latest/features/media-embed.html).
This plugin from CKEditor provides functionality for embedding various
types of media into the editor. If we want to use it again in the
future, we'll need to add Markdown syntax support for it, likely saving
it as various shortcodes supported by Hugo.

closes #576
pr #608