set up swc for jest, webpack
This sets up swc (the Speedy Web Compiler) for transpiling TS -> JS for
both webpack and Jest. swc (https://swc.rs) is a compiler for web assets
that supports bundling and whatnot. It's written in Rust, unlike most
frontend tooling like babel or the typescript compiler which are
written in JS and typescript, respectively. Being written in Rust means
that swc is significantly faster for a lot of tasks.

pr #861