add to validation schema, add an example thing in there
This adds a new option to the 'relation' field which is `sortable`. If
`sortable === true` then instead of using a normal react-select
multiselect input element, we show a more complicated UI with two
elements:

1) a single-select which acts like an 'add' input for selecting new
   pieces of content
2) a 'reordering' area which uses a drag-and-drop UI to allow the user
   to freely reorder the elements already selected.

Together this will make is easy for us to implement things like a video
gallery or similar.

closes #687
pr #719