add support for editing table shortcodes to the markdown editor
this adds a new markdown syntax plugin which represents tables (added
via the CKEditor plugin) with Hugo shortcodes instead of with normal GFM
syntax.

This syntax solves several issues with the GFM syntax, notably the lack
of support for line-breaks within table cells, related issues with
nesting things like lists, and so on.

Read issue #743 for details of why the shortcodes are set up the way
they are, how they are intended to work, etc.

closes #743
closes #673
closes #672
pr #744