add error handling to the Relation field request
The relation field makes a request to the site content API to fetch the
data that it formats as options for the user. Currently if this request
fails we have no error handling on it, so the whole app just crashes.

This request can fail because the site in question doesn't exist (#756)
or because of a permissions error (#738), or because of a more generic
network issue or whatever. In any case, this fixes it so that instead of
crashing the app we show an error message!

closes #759
pr #760