add cross_site option to the Relation field
This option allows us to create fields in our site configs which hold a
list of objects in other sites. If `cross_site` is set to `true` the UI
now will show a Select field for picking a website and then, once a
website is selected, it will show the normal multiple and sortable UI to
add items from that website to the list.

This overall will let us implement something like a resource list, where
we group, say, Math textbook resources from across all the courses on
OCW.

I also added a new ocw yaml file based on the one in ocw-hugo-projects
(by which I mean I copy-pasted). This has a cross-site relation field
implementing a resource list set up in it already, so if you set up a
new site which uses that configuration you should be able to test it
out.

closes #780
pr #813