remove Dockerfile-node
This is no longer necessary because we no longer need to do any specific
setup before our JS code is going to be able to run. Thanks yarn and
other new pieces of JS tooling!

So now we can just use one of the `node` images in `docker-compose.yml`
directly.

pr #831