Split sortable UI off from RelationWidget into SortableSelect component
I'm doing this in preparation for work we want to refactor course
collections. We want to add a new field type so that we can have course
collections (i.e. content author defined collections of courses, like
'Python-based programming courses', 'Math courses with video lectures,'
or something like that) be defined as content collections, like pages,
resources, and the like, which live in a site like ocw-www.

For this we could add more functionality to our existing RelationField
component, but I'm a little reluctant to do that because it's already
quite a large, complicated component. Instead I think it makes sense to
add a new WebsiteCollectionField field which will implement this
functionality.

In order to do that, I want to pull out the portion of the current
RelationField implementation which is concerned with reordering a list,
so that we can reuse that UI for the WebsiteCollectionField. This PR
basically does that, pulling that code out into a new SortableSelect
component which we'll be able to depend on for both things.

pr #869