add support for 'website-collection' field
This adds a new type of site content field called a `WebsiteCollection`.
This is designed to be a reorderable list of other websites in the
application, which can be used to support something along the lines of
the course collection feature that exists on current OCW prod. Under the
hood it uses the `SortableSelect` component I introduced in #869.

I also modified the checked-in ocw-www configuration yaml file to use
this new field in order to implement a 'Course Collection' content type.

closes #886
pr #888