remove 'legacy' implementation of WebsiteCollections
Before the new 'Website Collection' field type was added in #888 we had
a separate implementation of website collections which didn't save them
as course content but instead using separate `WebsiteCollection` and
`WebsiteCollectionItem` models. Since we're not going to continue with
that approach we might as well delete the code!

This commit removes both the frontend and the backend code related to
this feature.

closes #889
pr #890