switch from storing website UUID to the website name property
I thought initially for course collections that storing a website's uuid
would be what we need to then be able to render websites in a list in
ocw-www, but I goofed and missed that actually the name property is
what's going to end up being used in the URL to the data.json file for
each course at render-time, so that's actually the information we need.

Anyway - this PR just makes a small adjustment so that the website name
is saved in the course collection array rather than the uuid.

pr #895