add an optional filter to the website listing API for publish status
This adds a query param (published) to the website API which accepts a
boolean value and will filter accordingly on whether websites have
already been published or not.

It also makes a change to the WebsiteCollection widget to only show
published websites in it, ensuring that course content authors can't
accidentally create a course collection which references unpublished
courses.

pr #904
closes #887