fix up our handling of the camelcase eslint rule a little bit
Just a little neatness thing. For `types/websites.ts` we're declaring
the values we're getting from the backend, so we're stuck with camelcase
stuff there. I went ahead and disabled the `camelcase` rule for the
whole file, so we don't have to have a whole bunch of per-line disables.

Then I also made a few small changes to the `RelationField` widget to
improve the situation a bit.

pr #957