improvements to search handling on the Website listing API
This changes how we find results when the `search` parameter is passed
to the `WebsiteListing` API. Previously we were doing an `icontains`
filter on only the `title` field, which resulted in surprising behavior.
For instance, searching for course numbers would not return the relevant
courses, searching for the course `name` (the slug that goes into the
URL eventually) wouldn't either, and so on.

This changes things around a bit so that instead of a more basic filter
we leverage Django's support for Postgres' built-in full text search. It
is set up to search on the `title`, `name`, and `short_id` fields, which
should hopefully give us what we need to provide a search that works the
way course authors expect.

One shortcoming of this search support is that it does not support fuzzy
matching or stemming. Because of this, if you have a course called
'Philosophy' and you search for 'Philo' you won't get the results you'd
expect, and you'll instead have to (correctly spell and) type in the
whole word. This is a little unfortunate, but I think we can't fully
address it without moving to something more sophisticated, which is
probably more work than we want to do for a little filter in the UI like
this.

closes #990
pr #993