improvements to the site search
This fixes a few issues on the site search (the little box on `/sites`).
In particular, previously a full page refresh (well, actually a
react-router reboot of the whole route) was triggered when the search
changed. now the search results are updated, the search box retains
focus, and the UI around the results doesn't go away.

closes #963
pr #992