add support for nondestructive editing w/ legacy shortcodes
This implements support for nondestructively editing markdown files
which have 'legacy' shortcodes in them (these are shortcodes inserted
into converted content by ocw-to-hugo to support a series of things
which aren't supported in plain markdown).

A manifest of these shortcodes is declared in
`static/js/lib/ckeditor/plugins/constants.ts`, and support for them is
implemented in a new CKEditor plugin called `LegacyShortcodes`.

This will enable non-destructive editing of markdown content which
contains these shortcodes, fixing an issue we were seeing where raw
shortcodes that were unknown to CKEditor were being escaped and the
showing up in the rendered HTML. This fixes this by adding a node type
to CKEditor for each one, and adding basical logic for preserving any
arguments to the shortcodes.

Note that this change does NOT in any way way enable _editing these
shortcodes themselves_. Instead, they are made obvious in the editor
with a basic display, but it is up to the content editors to not delete
them or anything like that (and right now they could only be added to a
file manually in the Django admin or shell).

pr #1013