support resource link anchor IDs
We recently added support for an ID param on resource links
(mitodl/ocw-hugo-themes#444) in the course theme. We don't yet have
these in prod or anywhere, but soon we'll be converting some `{{<
baseurl >}}-based` links (#1036)  over to this format, so we want to
have support for it in CKEditor ready to go.

pr #1037