move website content drawer open / close / edit state to URL
This moves the state which controls the presence and contents of the
website content drawer into the URL bar. In particular, instead of
basically toggling a boolean and deciding whether to show the drawer or
not, instead we navigate to one of two new URLs:

| name         | url                                       |
| ------------ | ----------------------------------------- |
| edit content | /sites/:name/type/:contentType/edit/:uuid |
| new content  | /sites/:name/type/:contentType/new/       |

they do what you'd expect. I tried to do what I could to keep every
aspect of the UI the same except for the fact this this open / close
state is now driven by the URL.

closes #1030
pr #1072