fix spacing issue w/ single-line text inside of table cells
This fixes an issue we're seeing with single-line content in table
cells. Basically, the issue is that if you have markdown formatting on
that content it seems to be getting ignored by Hugo unless there is
white space around it, so something like this:

```md
{{< tdopen >}}**great stuff!**{{< tdclose >}}
```

_should_ render to

```html
<td><strong>great stuff!</strong></td>
```

but instead it's rendering to

```html
<td>**great stuff!**</td>
```

we can intervene in the html -> markdown conversion process and insert
newlines, like so:

```md
{{< tdopen >}}
**great stuff!**
{{< tdclose >}}
```

which fixes the issue!

closes #1082
pr #1081