add code block support to CKEditor
This adds [CKEditor's code block
feature](https://ckeditor.com/docs/ckeditor5/latest/features/code-blocks.html)
to our editor setup so that content authors can add and modify code
blocks.

They are saved in the normal way in Markdown with three backticks and
then the language name, like so:

    ```js
    const foo = "bar"
    ````

the nice thing about this is that Hugo already supports correctly
highlighting blocks which have the language set on them like this, so it
sort of just works!

closes #1083
pr #1088