fix a small issue with the website search
This changes the application of the `search` param on the website
listing API. Previously, it was filtering even if the search was `""`,
which led to no results being returned. This changes the behavior a
little bit so that filtering is only applied if the search string is
_not_ `""`.

This has a nice side effect that the `WebsiteCollectionWidget` will be
populated w/ options when the user first clicks in to it.

closes #1139
pr #1140