fix react-router / history typings
tldr: This fixes some issues we were having with react-router + history type definitions, and should ensure we don't break them again in the future.

Long run, it might be nice to upgrade to react-router 6 which is written natively in typescript so we won't have to futz around with DefinitelyTyped typedefs. But that might not be trivial to do.

Anyway...In a recent PR, I removed @types/history as a dependency of ours. We were actually using v5 and v5 of history is written in ts, so we didn't need the types. Right?

Wrong. It turns out that @types/react-router depends on @types/history, and its typedefs won't work without history typedefs. React router *says* it needs history ^4.7.11, but it actually needs max version 4.7.9. There was a breaking change between @types/history 4.7.9 -> 4.7.11 that broke how @types/react-router was using it.

I suspect the reason that @types/react-router never noticed was that the way in which it "broke" was that a whole bunch of things were being treated as "any" instead of their stricter types. So it didn't actually cause error messages, just ... because way less helpful.

Anyway, this ties us to v4.7.9 of @types/history.