do not emit declaration files (#1293)
tldr: Because we do not use them and they cause some ts errors in IDEs (that we suppress during CLI through `--noEmit`).

We currently have an error

> Exported variable 'getWebsiteDetailCursor' has or is using name '$CombinedState' from external module "/Users/chudzick/dev/ocw-studio/node_modules/redux/index" but cannot be named.

See https://github.com/microsoft/TypeScript/issues/5711 for some info about this error. Manually importing $CombinedState both seems silly and does not fix the issue (I think `re-select` would need to import it itself? I'm not sure. Possibly related to https://github.com/microsoft/TypeScript/issues/5711#issuecomment-231501080)

The error:
- is visible in some IDEs (e.g., VS Code)
- does not show up when running `npm run typecheck` since that uses the `--noEmit` CLI flag, which suppresses both compiled code and declarations from being emitted.

But our tsconfig does specify that declarations can be made...I'm not sure why. As I understand it, the setting makes sense for a library but not for an app.